cli: separate api command into own module
